+++ .vscode/settings.json
@@ -1,0 +1,9 @@
+{
+    "files.associations": {
+        "inet.h": "c",
+        "response.h": "c",
+        "io.h": "c",
+        "request.h": "c",
+        "in.h": "c"
+    }
+}
