+++ Makefile
@@ -1,4 +1,4 @@
-.PHONY: server runServer
+.PHONY: runServer
 
 OUT_DIR = target
 
@@ -8,4 +8,4 @@ runServer: $(OUT_DIR)/server/a.out
 	
 $(OUT_DIR)/server/a.out: src/server.c
 	mkdir -p $(OUT_DIR)/server
-	gcc src/server.c -o $@
+	gcc src/*.c -o $@
