--- index.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html lang="ja">
-  <head>
-    <meta charset="UTF-8">
-    <title>自作HttpServer</title>
-  </head>
-  <body>
-    <h1>Hello World!</h1>
-    <p>このページは自作HTTPサーバーから受信したページです</p>
-  </body>
-</html>
