+++ src/io.h
@@ -1,0 +1,6 @@
+#ifndef _IO_H
+#define _IO_H
+
+unsigned int getFileSize(const char *);
+
+#endif
