+++ src/response.h
@@ -1,0 +1,7 @@
+#ifndef _RESPONSE_H
+#define _RESPONSE_H
+
+int createResponseMessage(char *, int, char *, char *, unsigned int);
+int sendResponseMessage(int, char *, unsigned int);
+
+#endif
