+++ src/server.h
@@ -1,0 +1,2 @@
+int httpServer(int);
+void showMessage(char *message, unsigned int size);
